Value: split out the 'Variant' part
