+++ include/nix/Value.hpp
@@ -14,31 +14,15 @@
 #include <nix/DataType.hpp>
 #include <nix/Platform.hpp>
 #include <nix/None.hpp>
-
-#include <string>
-#include <cstdint>
-#include <stdexcept>
-#include <iostream>
+#include <nix/Variant.hpp>
 
 namespace nix {
+
 
 /**
  * @brief Class that corresponds to the odML Value.
  */
-class NIXAPI Value {
-private:
-    DataType dtype;
-
-    union {
-        bool        v_bool;
-        double      v_double;
-        uint32_t    v_uint32;
-        int32_t     v_int32;
-        uint64_t    v_uint64;
-        int64_t     v_int64;
-        char *      v_string;
-    };
-
+class NIXAPI Value : public Variant {
 public:
     double uncertainty = 0.0;
 
@@ -48,28 +32,19 @@ public:
     std::string checksum;
 
 
-    Value() : dtype(DataType::Nothing), v_bool(false) { }
+    Value() : Variant() { }
 
-    explicit Value(char *value) : dtype(DataType::Nothing) {
-        set(value);
-    }
+    explicit Value(char *value) : Variant(value) { }
 
-    explicit Value(const char *value) : dtype(DataType::Nothing) {
-        set(value);
-    }
+    explicit Value(const char *value) : Variant(value) { }
 
     template<typename T>
-    explicit Value(const T &value) : dtype(DataType::Nothing) {
-        set(value);
-    }
+    explicit Value(const T &value) : Variant(value) { }
 
     template<size_t N>
-    explicit Value(const char (&value)[N]) : dtype(DataType::Nothing) {
-        set(value, N);
-    }
+    explicit Value(const char (&value)[N]) : Variant(value) { }
 
-    Value(const Value &other) : Value() {
-        assign_variant_from(other);
+    Value(const Value &other) : Variant(static_cast<const Variant &>(other)) {
         uncertainty = other.uncertainty;
         checksum = other.checksum;
         encoder = other.encoder;
@@ -86,76 +61,17 @@ public:
         return *this;
     }
 
-    ~Value() {
-        maybe_deallocte_string();
-    }
-
-    void set(none_t);
-    void set(bool   value);
-    void set(int32_t value);
-    void set(uint32_t value);
-    void set(int64_t value);
-    void set(uint64_t value);
-    void set(double value);
-    void set(const char *value, const size_t len);
-    void set(const char *value);
-    void set(const std::string &value);
-
-    template<typename T>
-    T get() const {
-        T temp;
-        get(temp);
-        return temp;
-    }
-
-    void get(none_t &tag) const;
-    void get(bool &out) const;
-    void get(int32_t &value) const;
-    void get(uint32_t &value) const;
-    void get(int64_t &value) const;
-    void get(uint64_t &value) const;
-    void get(double &value) const;
-    void get(std::string &value) const;
-
-    DataType type() const {
-        return dtype;
-    }
-
     void swap(Value &other);
 
-    static bool supports_type(DataType dtype);
-
-private:
-
-    template<typename T>
-    void swap_helper(Value &other) {
-        T temp = get<T>();
-        assign_variant_from(other);
-        other.set(temp);
-    }
-
-    void assign_variant_from(const Value &other);
-    void maybe_deallocte_string();
-
-    inline void check_argument_type(DataType check) const {
-        if (dtype != check) {
-            throw std::invalid_argument("Incompatible DataType");
-        }
-    }
 };
-
-
-template<>
-inline const char * Value::get<const char *>() const {
-    check_argument_type(DataType::String);
-    return v_string;
-}
 
 
 NIXAPI std::ostream& operator<<(std::ostream &out, const Value &value);
 NIXAPI bool operator==(const Value &a, const Value &b);
 inline bool operator!=(const Value &a, const Value &b) { return !(a == b); }
 NIXAPI void swap(Value &a, Value &b);
+
+
 
 
 } // namespace nix
